Fixed double and int mistakes in BookDataController

+++ Backend/BackendCode/src/main/java/myProject/BookData.java
@@ -42,7 +42,7 @@ public class BookData {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(double rating) {
         this.rating = rating;
     }
